refactor: break out ModuleInfoCache from ParsedSourceCache
