+++ cli/cache/caches.rs
@@ -10,8 +10,8 @@ use super::cache_db::CacheDBConfiguration;
 use super::check::TYPE_CHECK_CACHE_DB;
 use super::deno_dir::DenoDirProvider;
 use super::incremental::INCREMENTAL_CACHE_DB;
+use super::module_info::MODULE_INFO_CACHE_DB;
 use super::node::NODE_ANALYSIS_CACHE_DB;
-use super::parsed_source::PARSED_SOURCE_CACHE_DB;
 
 pub struct Caches {
   dir_provider: Arc<DenoDirProvider>,
@@ -77,7 +77,7 @@ impl Caches {
   pub fn dep_analysis_db(&self) -> CacheDB {
     Self::make_db(
       &self.dep_analysis_db,
-      &PARSED_SOURCE_CACHE_DB,
+      &MODULE_INFO_CACHE_DB,
       self
         .dir_provider
         .get_or_create()
